pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nhi-security-testbed"
version = "0.1.0"
description = "Purple team framework for Non-Human Identity (NHI) security — attack simulation, detection validation, and SIEM-agnostic rule generation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "NHI Security Testbed Contributors" },
]
keywords = ["nhi", "machine-identity", "security", "siem", "wazuh", "purple-team", "detection-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core dependencies shared across mock services and agents
dependencies = [
    "flask>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "urllib3>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
# FastAPI-based API layer
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
]

# pySigma rule conversion pipeline
sigma = [
    "pySigma>=0.11.0",
    "pySigma-backend-splunk>=1.0.0",
    "pySigma-backend-microsoft365defender>=0.3.0",
    "pySigma-backend-elasticsearch>=1.0.0",
]

# Test suite
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.0",
    "pytest-timeout>=2.2.0",
    "pytest-xdist>=3.5.0",
    "docker>=6.1.0",
    "lxml>=4.9.0",
    "jsonschema>=4.20.0",
    "tenacity>=8.2.0",
]

# Development tools
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

# All optional dependencies combined
all = [
    "nhi-security-testbed[api,sigma,test,dev]",
]

[project.urls]
Repository = "https://github.com/your-org/machine-identity-discovery"
Documentation = "https://github.com/your-org/machine-identity-discovery/tree/main/docs"

[project.scripts]
nhi-demo = "scripts.run_demo:main"
nhi-health = "scripts.health_check:main"

[tool.hatch.build.targets.wheel]
packages = ["api", "src"]

# --- Tool configuration ---

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --timeout=60"
markers = [
    "smoke: Quick service availability checks",
    "integration: Cross-service integration tests",
    "e2e: End-to-end scenario tests (require running testbed)",
    "slow: Tests that take >30 seconds",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM"]
ignore = [
    "S101",  # assert usage OK in tests
    "S603",  # subprocess calls expected in testbed
    "S607",  # partial executable path OK for podman/docker
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]
"agents/**" = ["S104", "S105"]  # hardcoded demo secrets expected

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # gradual adoption
